requirements.txt
streamlit==1.26.0
Pillow